No more email/password support, combine /login and /join
